Fix type of exponent parameter

+++ CAENHVAsynApp/src/board_parameter.cpp
@@ -131,7 +131,7 @@ IBoardParameterNumeric::IBoardParameterNumeric(int h, std::size_t s, const std::
    if ( CAENHV_GetBdParamProp(handle, slot, param.c_str(), "Unit", &u ) != CAENHV_OK )
        throw std::runtime_error("CAENHV_GetBdParamProp failed: " + std::string(CAENHV_GetError(handle)));
 
-   int8_t e;
+   int16_t e;
    if ( CAENHV_GetBdParamProp(handle, slot, param.c_str(), "Exp", &e ) != CAENHV_OK )
        throw std::runtime_error("CAENHV_GetBdParamProp failed: " + std::string(CAENHV_GetError(handle)));
