+++ CAENHVAsynApp/src/channel_parameter.cpp
@@ -136,7 +136,7 @@ IChannelParameterNumeric::IChannelParameterNumeric(int h, std::size_t s, std::si
    if ( CAENHV_GetChParamProp(handle, slot, channel, param.c_str(), "Unit", &u ) != CAENHV_OK )
        throw std::runtime_error("CAENHV_GetBdParamProp failed: " + std::string(CAENHV_GetError(handle)));
 
-   int8_t e;
+   int16_t e;
    if ( CAENHV_GetChParamProp(handle, slot, channel, param.c_str(), "Exp", &e ) != CAENHV_OK )
        throw std::runtime_error("CAENHV_GetBdParamProp failed: " + std::string(CAENHV_GetError(handle)));
